+++ src/prop/bvminisat/core/Solver.cc
@@ -29,7 +29,7 @@ OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWA
 #include "base/output.h"
 #include "options/bv_options.h"
 #include "options/smt_options.h"
-#include "proof/bitvector_proof.h"
+#include "proof/resolution_bitvector_proof.h"
 #include "proof/clause_id.h"
 #include "proof/proof_manager.h"
 #include "proof/sat_proof.h"
@@ -1318,7 +1318,7 @@ void Solver::explain(Lit p, std::vector<Lit>& explanation) {
   }
 }
 
-void Solver::setProofLog( BitVectorProof * bvp ) {
+void Solver::setProofLog( ResolutionBitVectorProof * bvp ) {
   d_bvp = bvp;
   d_bvp->initSatProof(this);
   d_bvp->getSatProof()->registerTrueLit(mkLit(varTrue, false));
