+++ src/prop/sat_solver.h
@@ -32,7 +32,7 @@
 
 namespace CVC4 {
   
-class BitVectorProof;
+class ResolutionBitVectorProof;
 
 namespace prop {
 
@@ -97,7 +97,7 @@ public:
   /** Check if the solver is in an inconsistent state */
   virtual bool ok() const = 0;
   
-  virtual void setProofLog( BitVectorProof * bvp ) {}
+  virtual void setProofLog( ResolutionBitVectorProof * bvp ) {}
   
 };/* class SatSolver */
