+++ src/theory/bv/theory_bv.h
@@ -104,7 +104,7 @@ public:
 
   bool applyAbstraction(const std::vector<Node>& assertions, std::vector<Node>& new_assertions);
 
-  void setProofLog( BitVectorProof * bvp );
+  void setResolutionProofLog( ResolutionBitVectorProof * bvp );
 
 private:
